lesson6/pyproject.toml
[project]
name = "lesson6"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genai",
    "mcp>=1.6.0",
    "python-dotenv>=1.1.0",
]
